Test Code for a basic language Translator
I am following a tutorial Called "Let's Recreate Google Translate" by Edan Meyer on YouTube to get a good understanding of what I will be dealing with for this project. And since this is related to what our project is about, I'd like to see if this code will be helpful in any way.